+++ tests/testthat.R
@@ -1,4 +1,3 @@
 library(testthat)
-library(matsbyname)
 
 test_check("matsbyname")